be/app/requirements.txt
# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database (MongoDB) ---
motor==3.6.0              # Async driver cho MongoDB
pydantic==2.9.2           # Validate schema
python-dotenv==1.1.1

# --- Security (Auth) ---
python-jose[cryptography]==3.3.0   # JWT
passlib[bcrypt]==1.7.4             # Hash mật khẩu
python-multipart==0.0.9            # Hỗ trợ upload form/file (nếu login qua form-data)

# --- Config & Logging ---
pydantic-settings==2.4.0
loguru==0.7.2

# --- Development & Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx[http2]==0.27.0
